refactor[OK-135] Move board fetching functionality to a new file
